requirements.txt
jax
optax
equinox

tqdm

wandb

tensorflow-cpu
tensorflow-datasets
